Map tile redesign

+++ src/m_map.h
@@ -6,17 +6,17 @@
 // Map tiles are fixed in memory. The indentifer
 // for the data is not.
 struct MapTileData{
-	uint16_t tile_indx;
-	bool     is_passable;
-	int	 tile_texture_index;
-	int start_x;
-	int start_y;
-	int end_x;
-	int end_y;
+	int  tile_indx;
+	bool is_passable;
+	int  tile_texture_index;
+	int  start_x;
+	int  start_y;
+	int  end_x;
+	int  end_y;
 };
 struct MapData{
-	uint8_t  tileset;
-	uint32_t count;
+	int tileset;
+	int count;
 	struct MapTileData data[];
 };
 #pragma pack(pop)
